// src/utils/enums.ts
export enum Size {
  XS = 'xs',
  SM = 'sm',
  MD = 'md',
  LG = 'lg',
  XL = 'xl',
  XXL = 'xxl',
  FULL = 'full',
}

export enum Spacing {
  XXS = 'xxs',
  XS = 'xs',
  SM = 'sm',
  MD = 'md',
  LG = 'lg',
  XL = 'xl',
  XXL = 'xxl',
}

export enum IconName {
  ARROW_BACK = 'arrow-back',
  ARROW_RIGHT = 'arrow-right',
  CHAPTER = 'chapter',
  CHECKMARK = 'checkmark',
  CHEVRON_DOWN = 'chevron-down',
  CHEVRON_LEFT = 'chevron-left',
  CHEVRON_RIGHT = 'chevron-right',
  CLOSE = 'close',
  DOWNLOAD = 'download',
  DOWNLOAD_ALT = 'download-alt',
  DROPDOWN = 'dropdown',
  EMAIL = 'email',
  EMAIL_ALT = 'email-alt',
  EXTERNAL = 'external',
  FACEBOOK = 'facebook',
  FACEBOOK_ALT = 'facebook-alt',
  FACEBOOK_SQUARE = 'facebook-square',
  FILTER = 'filter',
  GRID = '2x2grid',
  HAMBURGER_MENU = 'hamburger-menu',
  HAMBURGER_FOUR = 'hamburger-four',
  HOME = 'home',
  INFORMATION = 'information',
  INSTAGRAM = 'instagram',
  LINKEDIN = 'linkedin',
  LINKEDIN_ALT = 'linkedin-alt',
  MINUS = 'minus',
  MINUS_CIRCLE = 'minus-circle',
  PLAY = 'play',
  PLUS = 'plus',
  PLUS_CIRCLE = 'plus-circle',
  REDDIT = 'reddit',
  RSS = 'rss',
  SHARE = 'share',
  SEARCH = 'search',
  TWITTER = 'twitter',
  TWITTER_ALT = 'twitter-alt',
  TWITTER_SQUARE = 'twitter-square',
  UNION = 'union',
  VIDEO = 'video',
  YOUTUBE = 'youtube',
  GITHUB = 'github',
  MICROPHONE = 'microphone',
}

export enum IconColor {
  'PRIMARY-C' = 'primary-c',
  BLACK = 'black',
  'PRIMARY-A' = 'primary-a',
  BLUE = 'blue',
  CURRENT = 'current',
  'PRIMARY-B' = 'primary-b',
  'PRIMARY-B-500' = 'primary-b-500',
  GRAY = 'gray',
  TRANSPARENT = 'transparent',
  'PRIMARY-D' = 'primary-d',
  WHITE = 'white',
}

export enum TextStyles {
  CAPTION = 'caption',
  DATE = 'date',
  DISPLAY = 'display',
  EYEBROW = 'eyebrow',
  EYEBROW_WITH_BACKGROUND = 'eyebrow-with-background',
  SUBTITLE = 'subtitle',
  PARAGRAPH = 'paragraph',
  SMALL = 'small',
  HERO_TEXT = 'hero-text',
  HEADING1 = 'heading-1',
  HEADING2 = 'heading-2',
  HEADING3 = 'heading-3',
  HEADING4 = 'heading-4',
  HEADING5 = 'heading-5',
  HEADING6 = 'heading-6',
  SIDE_HEADING = 'side-heading',
  CARD_HEADING = 'card-heading',
  HIGHLIGHT_EYEBROW = 'highlight-eyebrow',
  BANNER = 'banner',
}

export enum StripeColor {
  STRIPE_LIGHT_PRIMARY_A = 'stripe-light-primary-a',
  STRIPE_DARK_PRIMARY_A = 'stripe-dark-primary-a',
  STRIPE_TERTIARY_A = 'stripe-tertiary-a',
  STRIPE_SECONDARY_B = 'stripe-secondary-b',
  STRIPE_NEUTRAL = 'stripe-neutral',
  STRIPE_LIGHT_BLUE = 'stripe-light-blue',
  STRIPE_DARK_BLUE = 'stripe-dark-blue',
  STRIPE_GREEN = 'stripe-green',
  STRIPE_YELLOW = 'stripe-yellow',
  STRIPE_GRAY = 'stripe-gray',
}

export enum ButtonVariants {
  BTN_PRIMARY = 'primary',
  BTN_PRIMARY_BLUE = 'primary-blue',
  BTN_PRIMARY_BLUE_DARK = 'primary-blue-dark',
  BTN_PRIMARY_DARK = 'primary-dark',
  BTN_SECONDARY = 'secondary',
  BTN_SECONDARY_BLUE = 'secondary-blue',
  BTN_SECONDARY_DARK = 'secondary-dark',
  BTN_DROPDOWN = 'dropdown',
  BTN_SECONDARY_TRANSPARENT = 'secondary-transparent',
  BTN_TERTIARY = 'tertiary',
  BTN_DISABLED = 'disabled',
  BTN_SECONDARY_NO_BORDER = 'secondary-no-border',
  BTN_WHITE = 'white',
  BTN_WHITE_DARK = 'white-dark',
  BTN_NAV_LINK = 'nav-link',
  BTN_MOBILE_NAV_LINK = 'mobile-nav-link',
}