// src/components/lbj-button/lbj-button-types.ts
export type ButtonVariant =
  | 'primary'
  | 'primary-blue'
  | 'primary-blue-dark'
  | 'secondary'
  | 'secondary-blue'
  | 'primary-dark'
  | 'secondary-dark'
  | 'dropdown'
  | 'secondary-transparent'
  | 'secondary-no-border'
  | 'nav-link'
  | 'mobile-nav-link'
  | 'white'
  | 'white-dark'
  | 'tertiary';